test(Header, News): adicionar testes
