+++ src/templates/News/News.test.tsx
@@ -1,0 +1,27 @@
+import { ThemeContextProvider } from '@/Contexts/ThemeContext/ThemeContext'
+import { render, screen } from '@testing-library/react'
+import { News } from '.'
+
+const renderComponent = () => (
+  render(
+    <ThemeContextProvider>
+      <News />
+    </ThemeContextProvider>
+  )
+)
+
+describe('News', () => {
+  const useRouter = jest.spyOn(require('next/router'), 'useRouter')
+  useRouter.mockImplementationOnce(() => ({
+    back: jest.fn()
+  }))
+
+  it('should render the page with its components', () => {
+    renderComponent()
+    screen.getByTestId('news_header')
+    screen.getByTestId('news_banner')
+    screen.getByTestId('news_details')
+    screen.getByTestId('news_article')
+    screen.getByTestId('news_scrollBackToTopButton')
+  })
+})
